Clearly separate test “phases” from test “types”.
A test “phase” relates to the state that the Grails application/infrastructure is in at the time of test execution. A test “type” maps to a directory in the tests folder and represents a style of test.

When a “phase” starts, it calls “«phase»TestPhasePreparation” to set up the state (for example, “functionalTestPhasePreparation()” starts up the Grails app and embedded container). Each “type” registered for that “phase” is then executed. Before each test “type” is executed, “«type»TestsPreparation()” is called. This _must_ return a “GrailsTestHelper”. After the tests have been run, “«type»TestsCleanUp()” is called. After all of the “types” for a “phase” have been executed, “«phase»TestPhaseCleanUp” is called.

This new approach makes it much easier to implement new test “types”. It also has the advantage of making it easier for test “types” to participate in existing “phases” (e.g. functional-test and selenium plugin can run in the one “functional” phase and avoiding starting/stopping the app twice).