Add support for ContextResource
Some Resource implementations don't have a path property, but implement org.springframework.core.io.ContextResource, which has a pathWithinContext property instead.

This change is required in order to run a Grails app in an OSGi environment, as the Resource is usually a org.springframework.osgi.io.OsgiBundleResource instead of a org.springframework.core.io.UrlResource.