GRAILS-6159 - when autowiring tests, explicitly process them with an @Autowired annotation post processor.
The autowireBeanProperties() method gets called for a new instance of the test class for each test method. Previously, only the first call to this for any instance of a particular class was populating fields annotated with @Autowired (only if the autowire mode is AUTOWIRE_BY_NAME which it needs to be). Subsequent calls were failing to autowire the annotated properties. It is unclear if this is a bug or intentional behaviour from Spring.

The solution is to explicitly process beans with a AutowiredAnnotationBeanPostProcessor _after_ being autowired by the bean factory.