GRAILS-5863 - manually issue redirects instead of using sendRedirect() to avoid committing the response at the time the redirect is issued.
To compensate for this, we augment the getRedirectedUrl() method of the mock response we use for testing to look at the status code and “Location” header in GrailsMockHttpServletResponse.

A side affect of this is that  we had to change both some internal Grails tests and the user testing infrastructure to use GrailsMockHttpServletResponse over Spring's MockHttpServletResponse so that redirects can be tested for (via response.redirectedUrl) as they have been up until now.