GRAILS-8017 - Improve handling of private fields added via AST transformations
When adding the api instance property AbstractGrailsArtefactTransformer
was checking to see if a field already exists with the corresponding
name and if it did, no new field was added.  A problem with that is that
if the field was declared in a parent class and was private, the child
class would not have access to it.  This change improves that handling
to not only check to see if a field already exists but if it does and it
is private and it is declared in a parent class then we go ahead and add
a new private field with the same name to the class being transformed.