GRAILS-7979 - Improve Command Object Reloading
The constrainedProperties property that is added to validateable classes (like command object classes) is static. When the reloading agent reloads a class that property needs to be nulled out so that it will be re-initialized, in case the reload was initiated by changing constraints.  This commit adds a statement to the validateable classes static initializer to null that property out.  Grails runs the static initializer when classes are reloaded by the agent.