GRAILS-8319 - Fix grails.dbconsole.enabled
The grails.dbconsole.enabled Config.groovy property is supposed to allow enabling and disabling of the H2 db console.  The dbconsole is enabled via a servlet added to web.xml by DataSourceGrailsPlugin.doWithWebDescriptor.  A problem is that doWithWebDescriptor was only being run under certain circumstances so if the Config.groovy property was changed, until web.xml was regenerated the property change would have no effect.  Now we are running doWithWebDescriptor unconditionally.