GRAILS-8282 - add checkDuplicateDependencies task
The task will check the dependency graph for multiple versions of the same dependency and fail the build if any are found.  For example, if the build depends on ant 1.7 and ant 1.8 the build will fail.  The check depends on the module and group names being exactly the same.  Duplicate dependencies that don't have the same module and group will not be identified.