Don't filter project dependencies out of the configurations used to populate the lib dir.
Filtering out the project dependencies was causing the configuration to resolve differently due to conflicts introduced by project dependencies not being taken into account. Instead, we filter when copying the configurations assuming that any dependency with a module name that matches the project name of any projects in our build is a project dependency.