GRAILS-8405 - Problem with resource tag
NOTE: This is not a fix, but a temporary change.

There seems to a be a problem where the wrong Writer is being used in certain circumstances.  One problematic scenario is if a default codec is in place, the resource tag is used in the HEAD and then used again later in the BODY, content seems to be written the wrong part of the response.  This change avoids the problem but circumvents some intended optimizations.  This will be revisited, soon.

A relevant functional test has been added: https://github.com/grails/grails-functional-tests/commit/0cdcaeeec236f1978a4a0711900b227085e4a935