GRAILS-8059 - Support relocating the spring loaded cache dir
This is incomplete but is a start for supporting the relocation of the spring loaded cache directory by assigning a value to the GRAILS_SL_CACHE_DIR environment variable.  The name of the variable and the way we handle this is subject to change.  Note this change only relates to the unix startup scripts.  In particular, this does not include any support for Windows yet.