Fix for GRAILS-8700
Cache key collision was caused by the default toString of `ByteArrayResource`. This meant if multiple calls to `g:render` for a mocked view were made in a single unit test the template for the first call was reused for all subsequent calls as the cache key for any mock GSP resource was always the same.