GRAILS-8463: use type="text" for decimal inputs
Because of localization issues when using `type="number"` for decimal properties I've reverted to using `type="text"` and formatting the value for the locale.

The only downside I can see is that the scale is ignored by `g:fieldValue` whereas I think it would be nice if it was used to render the appropriate number of decimal places. That's really a separate issue, though.