GRAILS-9509 - improve reloading
There is a problem with stale JVM state that isn't cleared up properly on reload.  Spring Loaded is going to address this but for now the simplest thing for us to do is the change represented here in this commit.

See comments at http://jira.grails.org/browse/GRAILS-9509 for more info.