GRAILS-9674 - fix GrailsEnvironment
The AbstractEnvironment constructor is calling customizePropertySources.  Our GrailsEnvironment class was overriding this method with code that referred to grailsApplication, which will not be initialized in time because it is initialized in the GrailsEnvironment constructor, which of course is called after the parent class constructor completes.  This rearranges the code to try and address this.  We need to test this and verify that all of the environment support stuff still works with this change.