Allow reloading with MingW bash on Windows.
Fixes GRAILS-8271. It seems that Java on Windows doesn't like the path to the
reloading agent being Unix style. This hack converts the path to Windows style
before its added to the launch command for Grails.

This commit fixes the problem with the earlier one where an error was logged to
the console when running on non-Windows platform. The `startGrails` script now
has a `$mingw` variable that is only set to true if the platform `uname` is the
MingW one.