GRAILS-10753 Console Colors in Windows
The GrailsConsole makes sure now that whenever ANSI escape codes are output to the console they are followed by the reset sequence, so that the console is not 'stateful' when it comes to the colouring, formatting, etc. It applies to the userInput as well, so that even if the prompt question is coloured the reset mark precedes the user's input. It ensures that if the process is killed in-between, the console is set to the default state.