Improve data binding tests
Retrieve the data binder from the application context which will make it easier to test built in converters because the binder will be configured via DataBindingGrailsPlugin.  This will help test issues like GRAILS-11105.

Some tests needed to be pulled out of GrailsWebDataBinderSpec and put into GrailsWebDataBinderConfigurationSpec because they are mutating the binder which can cause problems for other tests now that the binder is being pulled from the context.