GRAILS-10090 - improve view rendering for namespaced controllers
The default view path for namespaced controllers will now be grails-app/views/controllerNamespace/controller/view and if that does not exist the fallback is the previous default of grails-app/views/controller/view.

Squashing commits from local branch to eliminate the evolution and later elimination of complexity.  This ends up being fairly simple.