Make sure any transitive Spring dependencies use the correct version of Spring and don't pull in older versions which can break Grails.
This is done by making Spring managed dependencies at the Aether level. Ideally it would be nice to allow the user to specify managed dependencies, but we don't have an API for this yet.

Note that it is still possible for the user to use an older version of Spring if they declare a dependency at the top level (depth == 0)