GRAILS-11661 - unit test metaclass cleanup issue
This test uses @FreshRuntime to satsify the requirement.  I don't know if we are going to be able to support cleaning up changes between test methods.  One problem with that is trying to roll back changes made while a test method is running that overlap with changes made while setting up the test environment.  When rolling back the changes made while the test method is running, changes made while setting up the test environment would also be rolled back.  We need to decide if it is worth introducing the complexity necessary to support rolling back changes made while test methods are run.  It may be that we just recommend @FreshRuntime for those scenarios.  Pending discussion with the rest of the team.