GRAILS-11673 - improve unit testability of the forward method
This is a breaking change for existing tests which expect the leading "/grails" and trailing ".dispatch" to be there but this change makes the system make more sense.  It is only a breaking change for unit tests and should not affect any running applications.  We will note the change in the release notes.