fix failing tests
Related to 0a31401e2e16257ecd146a771e94d568b78eae21

Some of this is temporary.  Part of the issue is that we have tests which do dynamic compilation of classes and if we compile a class with the same name twice, that is problematic because of changes in the commit mentioned above.  The need to keep track of which classes have been compiled is a temporary workaround pending a real solution.  Classes need to have unique names for now.