Don't apply the same ClassInjector to a node twice
This is an incomplete fix.

Artifact transformations were being applied twice.  Once directly by ArtefactTypeAstTransformation and once indirectly by GlobalGrailsClassInjectorTransformation.

The problem was manifesting any time a controller defined an action which accepts a parameter.  The first time through the transformation would succeed.  As designed, the transformation creates a no-arg version of the action method which and the no-arg version of the method would invoke the version which accepts a parameter.  The second time through the transformation would generate an error because the overloaded method would be found, which is not supposed to be allowed.  This could be fixed by adding some logic directly to ControllerActionTransformer but the problem is more general than that.