commenting out test method
This test method is failing with Groovy 2.4.0-SNAPSHOT.  It appears that the relevant behavior is working fine and there is something wrong in the test suite.

The test passes when run in isolation and passes when run along with other test methods in the same test class but is failing when run along with other test classes.

The following passes:

./gradlew :g-t-s-web:singleTest -DsingleTest.single=*FormTagLibTests*

The following fails because it will include not only FormTagLibTests but also FormTagLib2Tests and FormTagLib3Tests:

./gradlew :g-t-s-web:singleTest -DsingleTest.single=*FormTagLib*

I haven't pinned down the specifics yet but it looks like the problem may be causing the mapping to not be found in the mappings holder.  Pending investigation.