Fixes #693 by ensuring maven is used in forked mode for inline plugins
This additional check for externally configured dependencies was added with this commit: 40c3c1e13fe459d57fcbb97c64f6491a37c6dce5.  The issue is that when forked, isDependenciesExternallyConfigured is true and therefore Ivy is used when Aether should be used for inline plugins. Removing this check fixes the issue and there did not appear to be any side effects of configuring a dependency manager even if dependencies were configured externally.